fix: harden JSON parsing and unify schemas
